add/component-graphic
